pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-cli"
version = "0.1.0"
description = "CLI tools for Claude Agent Framework"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Claude Agent Framework" }
]
dependencies = [
    "typer[all]>=0.9.0",
    "duckdb>=0.10.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
caf = "claude_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_cli"]

[tool.hatch.build.targets.sdist]
include = ["src/claude_cli"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
